reduce header dependencies
